+++ app/src/main/res/layout/activity_main_screen.xml
@@ -44,14 +44,6 @@
         android:layout_height="match_parent"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"/>
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_marginEnd="16dp"
-        android:layout_marginBottom="16dp"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:src="@drawable/mscreen_btn_add"
-        android:backgroundTint="@color/marigold"
-        android:layout_gravity="bottom|end"/>
+
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
