+++ app/src/main/res/layout/fragment_budget.xml
@@ -20,7 +20,15 @@
     </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
 
 
-
+    <com.google.android.material.floatingactionbutton.FloatingActionButton
+        android:id="@+id/fab"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="16dp"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:src="@drawable/mscreen_btn_add"
+        android:backgroundTint="@color/marigold"
+        android:layout_gravity="bottom|end"/>
 
 
 </FrameLayout>
